requirements.txt
brotlipy==0.7.0
certifi
cffi
charset-normalizer
click
cryptography
Flask
idna
itsdangerous
Jinja2
MarkupSafe
pycparser
pyOpenSSL
PySocks
python-dotenv
requests
urllib3
Werkzeug
